pyproject.toml
[project]
name = "route-pathfinding-visualizer"
version = "0.1.0"
description = "Dual-map route visualizer comparing A* and Dijkstra pathfinding algorithms"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.30.0",
    "folium>=0.15.0",
    "geopy>=2.4.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "streamlit-folium>=0.15.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pylint>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pylint.main]
fail-under = 8.5
max-line-length = 100

[tool.pylint.messages_control]
disable = [
    "C0330",  # Wrong hanging indentation (conflicts with black)
    "C0326",  # Bad whitespace (conflicts with black)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests requiring API calls",
    "benchmark: Performance benchmark tests",
]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
]
